taint/test/mochitest/test_window.html

<!DOCTYPE HTML>
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <meta charset="utf-8">
    <title>Check Window</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script "text/javascript">

      function check_tainted(str) {
          ok(str.taint.length > 0, "Check tainted");
      }

      function check_untainted(str) {
          ok(str.taint.length == 0, "Check untainted");
      }

      add_task(async function test_window_name() {
        window.name = "tainted";
        check_tainted(window.name);
      });
    </script>
  </head>

  <body/>

</html>
